Refactor PDA helpers for consistency
