+++ clients/js/src/hooked/index.ts
@@ -1,6 +1,6 @@
 export * from './changeLog';
 export * from './concurrentMerkleTree';
 export * from './merkleTreeAccountData';
-export * from './mintAuthorityPda';
+export * from './mintAuthority';
 export * from './path';
 export * from './resolvers';
